requirements.txt
numpy
pillow
tqdm
tensorflowjs
pyinstaller
tf2onnx
onnxruntime